requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pillow==10.0.1
ultralytics==8.0.196
python-multipart==0.0.6
streamlit==1.27.2
opencv-python-headless==4.8.1.78
matplotlib==3.8.0
numpy==1.26.1
requests==2.31.0
pytest==7.4.3
python-dotenv==1.0.0
